refactor(web): split training-plan.store actions into sibling modules
Behaviour-preserving split of training-plan.store.ts (602 → 227 LOC) to meet
the ≤250 LOC rule. The store stays a thin orchestrator (props + computed +
one-line method wrappers); the action/effect bodies move to focused modules:

- training-plan-store.actions.ts — in-flight-lock + reps helpers and the
  day-logging/completion actions (logPlanDay, markTodayDone, markDayDone, …)
- training-plan-store.lifecycle.ts — plan-lifecycle actions (start, skip/unskip,
  jumpToDay, unmarkDayDone, abandon)
- training-plan-store.hooks.ts — the browser-only day-tick interval + auto-mark
  effect, registered from withHooks.onInit

Public store API and behaviour are unchanged (LogPlanDayResult re-exported);
110 training-plans tests pass.

Refs #456
